Remove old database systems and old bin.clj
These all need to be ported over to the new jepsen.core system, but
folks are still trying to run them on master. If you're looking for
these tests, see the branch called "old". If you're feeling ambitious,
help port these tests from the old branch to master, following the style
of jepsen.system.etcd, and so on.