Support etcd v2.0
Newer versions of etcd have deprecated many essential command line
options for setting up a static cluster [1]. This caused exceptions
on the cluster machines and also in the Jepsen script. This patch
tweaks things so that the cluster will start up with etcd v2.0.

Note that the LXC instructions for setting up the host names are
slightly adjusted.

[1] https://github.com/coreos/etcd/blob/master/Documentation/clustering.md#04-to-20-migration-guide