util: Rewrite timeout, remove unwrap-exception
future-cancel will call Thread.interrupt, so we don't need to do it
here (in fact I think it's racy to call it here: the thread could have
moved on to other work and we could be interrupting something else.
future-cancel knows to only interrupt the thread if it's still working
on the future.

defer may raise either ExecutionException (which has a cause) or
InterruptedException (which does not). unwrap-exception would
transform the latter into an unhelpful NullPointerException.
Replace unwrap-exception (which was used nowhere else) with a more
targeted try/catch.