Configure "good enough" replication.
If there is a single replica, place all process in that replica, and
set the primary to be the single log node. If there is more than one
replica, mod partition processes into each replica and each log
partition, such that replicas == log partitions.

Note that if the number of processes is not evenly divisable by the
number of replicas, one replica will not have all log partitions.