Fix a deadlock caused by auto/start!
A thread starting the DB while another thread shut it down shortly
thereafter could spin indefinitely in a single command, waiting for the
DB to bind ports, which would prevent Jepsen's teardown process from
closing the SSH connection, since the connection was never not in use.
This change allows us to leave the SSH connection occasionally so we can
tear it down.