jepsen.checker.perf: improvements to nemesis graphs
This creates a new abstraction: a plot data structure, and creates a
function, plot!, to plot that structure. We refactor nemesis-related
plotting out of each of the plotting functions, and move it to a shared
function that transforms plot objects. This should make it easier to add
nemesis operations to other types of plots later.

This also solves two practical problems:

1. We only render nemesis legend elements for those nemeses that
actually participated in the history, rather than *every* nemesis.

2. The color and shading of nemesis operations is much closer, instead
of rendering a black bar in the legend and a light grey one on the plot.