jepsen.tests.long-fork: Fix a serious bug where we failed to find forks
The long fork checker failed, in most cases, to notice long forks.

It was originally designed to work with the same order of keys in every
read in a group; however, we later randomized the order. This prevented
the long-fork checker from realizing that those reads actually involved
the same set of keys; it assumed every order was a distinct group with
(typically) only a single read of that group, and single reads trivially
pass, since there are no other reads to conflict with.

We now construct an order-independent *set*, rather than a *sequence*,
of keys observed by each read, and use that set to find which reads
belong to the same group.