jepsen.control: Add a `Controller` protocol
Not everywhere can use SSH, or can use SSH with the built-in
`clj-ssh`. To support these sorts of environments we add a protocol
called `Controller`, which provides the core abstractions that the
rest of the control facilities need: connect and disconnect sessions,
upload and download files, and run remote commands. (It's kinda like
`Client`, only it's for control-plane operations, not data-plane
operations.)