nemesis.membership: A state machine framework for membership changes.
This has historically been one of the hardest things to write in
tests--you have to figure out what the cluster looks like, what it
*should* look like, and also what every node individually thinks the
cluster should look like, and also what sort of transitions might be in
progress---understanding that those operations could take a long time to
resolve, and might need to be interleaved with other nemesis operations,
like resolving partitions, to complete---and emitting operations which
are appropriate for this state, and then *applying* those operations to
the cluster, and keeping track of the fact that you did those things.
Also there are performance implications to querying the cluster state,
so you want to cache that, but also probably update it asynchronously,
which means you're having to manage future lifecycles, and do error
handling, etc etc.

nemesis.membership tries to provide a framework for this awful mess so
you can focus on logical questions like:

- What does this node think the cluster state is?
- How should I combine the views from different nodes?
- What operations are possible right now?
- How do I apply an operation?
- How do I know when an operation is done?

I've written a half-dozen ad-hoc implementations of membership nemeses,
and built this abstraction with those systems in mind. It's been
developed for a new test, but I'm still not at all sure it's the right
shape for the problem in general. APIs and semantics are probably
going to change over time, so treat this as experimental.