jepsen.control: separate *password* and *sudo-password*
Jepsen no longer provides your password to `sudo` commands by default.
Set :sudo-password in your test's :ssh map to enable this, if necessary.

I've *really* tried to avoid this, but I cannot figure out a reliable
way to control whether sudo actually prompts for a password. If you're
running as root, or using other accounts with NOPASSWD (which is
common), then sudo *won't* need a password on stdin--which means we
wound up putting the user's password into the stdin of whatever command
was being run. That hasn't *typically* bitten us, but a few Jepsen
commands use stdin now, and that did not go well!