control/env: nicer construction of env var strings
We're doing a little more work with environment vars these days, and
start-stop-daemon has this sort of hacky `:env` option which was
supposed to be a string or series of strings. However, if those strings
triggered shell escaping, things got *weird*: it was hard to remember
exactly what to feed it. This commmit adds a new function, control/env,
for constructing env var strings out of literals, strings, and maps. You
can now pass maps like {:AUTHOR "Foucault"} as the :env for
start-daemon. Hopefully this is a somewhat smaller footgun.