nemesis.membership: add State setup/teardown, fix docs
nemesis.membership.States now have setup! and teardown! callbacks which
can be used to squirrel away state (e.g. network clients) that may be
needed for interacting with the cluster.

The docstring for nemesis.membership/package had nothing to do with how
it actually worked: it was for an earlier design of the membership
system which was rewritten. Now it correctly documents behavior, which should
make it *much* easier to use.