generator: stagger & delay changes, new cycle-times gen
cycle-times solves a frequent problem in Jepsen generators: you want to
alternate between two (or more) generators on a fixed timescale. For
example, you might have a generator of nemesis faults, and every 60
seconds, take 30 seconds off to allow the cluster to recover.
cycle-times preserves the state of each generator through this
alternation, which means that generator invariants like "we never
perform two start ops in a row" are preserved as well.

stagger and delay are slightly changed: they used to "catch up"
aggressively when they couldn't keep up with request rates. This caused
problems: for one, when a database went through a period of timeouts it
could cause a stagger/delay to become significantly behind schedule, and
that generator would then spam the DB with tons of requests once it
recovered. Second, it played poorly with cycle-times, which
*intentionally* leaves a generator quiet for some time.

Now, stagger and delay attempt to emit operations whose invocations are
separated by roughly dt seconds, *without* trying to catch up when they
fall behind.