store.format: Make block indexes an appendable block
Instead of preallocating a block index and overwriting it in place, we
append block index blocks to the end of the file like regular blocks,
and store the offset of the main block index at the start of the file
instead. This adds an extra disk seek during parsing, but simplifies
writing and provides, I hope, an extra measure of crash safety; we can
compact/rewrite later for read efficiency if necessary.

Also introduces a unified function for reading all types of blocks and
parsing their contents.