nemesis.membership: invoke! can return [op', state'] pairs.
This allows membership records to record state changes resulting from
invocations--for example, recording that a node is being removed before
issuing a remove-node command to the cluster.

As a part of this change, we now acquire locks on any changes to the
membership state. This shouldn't be a huge deal--it only changes due to
folding in new membership views and calls to invoke! We block for
the duration of invoke!, but that's fine because invoke! is already
single-threaded. We still fetch node views asynchronously, but merging
them in has to wait for the state lock.