generator.interpreter: ops can have :end-process? true
If a client returns a completion op with `:end-process? true`, it
forces the interpreter to end that process *even if it did not crash*
and start a new one.

This is helpful for network clients, like Kafka, which completely
explode once anything bad happens to them, and have to be torn down and
restarted from scratch, even though you can *prove* the operation you
were doing didn't take place, which means it's :fail, not :info.