db/Process is now db/Kill; should also work with AOT compiled deps
The weird eval hack we were doing to define a protocol called Process
(conflicting with java.lang.Process) resulted in (in AOT-compiled tests)
multiple copies of the Process protocol, and caused weird, silent
failures in test suites. In particular, you might run a test which was
supposed to use nemesis.combined to kill processes, and it would (seeing
that the DB you provided implemented the wrong version of Process)
refuse to actually kill anything.

We now resolve this issue with a sneaky double-eval hack: both AOT and
non-AOT dependencies should work.

Since all this seems fragile, I've also renamed jepsen.db/Process to
jepsen.db/Kill, which is symmetric with the Pause protocol. Tests using
Process should continue to work just fine, but we can migrate to Kill
over time and eventually dispose of this hack.