store.format: BigVector is now a full persistent vector
This should let us be compatible with (I hope) any checker that expected
histories to be real vectors. Most operations fall back to materializing
the whole vector in memory and then mutating that, under the assumption
that if you're assoc'ing *one* index, you're probably assoc'ing *lots*
of indices, and disk backing is probably moot.