cli: analyze no longer needs you to re-pass the original test args
The `analyze` CLI command now extracts the original argv from the
test stored on disk and uses it to re-inflate a test map. Now you can
run `lein run analyze -t -1` or `-t store/foo/1234...Z`, and it'll
figure out the rest. This should make it a *lot* easier to re-analyze
tests.

This is a breaking API change technically, but analyze has a.) always
been fragile, and b.) because of that really only gets used
interactively, and I'm hoping the new style is so much easier that folks
will forgive the change.