generator.context: fast, generic contexts for generators
Generators do a *lot* of context mutation and lookup. These operations
have been expensive and difficult to optimize because of the explicit
representation used by generators: a map with `:workers`,
`:free-threads`, etc. They were also a mix of Bifurcan sets and Clojure
collections.

This makes generators a defrecord: they'll still work fine with assoc,
which is how they're intended to be used, but the :workers and
:free-threads state are gone. In their place is a new interface (and
wrapper functions) for contexts. All references to direct fields (other
than :time) in contexts have been replaced with these functions. This
frees up those functions to do smarter bookkeeping, and in general for
us to experiment with other data layouts later.

This *might* be a breaking API change, but I don't think hardly anybody
was actually interfacing with these fields in context maps.