generator: don't generate Ops with duplicate keys in __extmap
This was... truly ridiculous. We generated Op defrecords which had not
only :value :foo, but *also* an extmap including {:value :foo}. Under
certain conditions these two values could diverge, creating absolutely
nightmarish bugs. This was due to a trivial bug in gen/fill-in-op which
used the full initial map as the extmap, rather than dissociating the
standard keys as intended.