basis->proposed (#21)
* Quiet down logging during tests

* checker/check-safe: write errors to results.edn using datafy, not a horrible string.

* Version 0.3.6

* Version 0.3.7-SNAPSHOT

* jepsen.history 0.1.4

* Version 0.3.7

* Version 0.3.8-SNAPSHOT

* Fix catch with throw+ instead of throw

The catch is in `try+`, which can catch other things than throwable. So
use `throw+` to re-throw the exceptions.

* web: shell out for zip files

We can get a little extra speed by calling out to /usr/bin/zip for
constructing zip archives of test directories.

* Checker.perf: configurable filenames for latency-raw.png

* web: catch errors calling out to /usr/bin/zip properly

* docker: install JDK manually

* docker: remove apt-preferences

* nemesis.time/bump-gen-select: flatten time distribution

With an exponential factor of 2, most of the test was spent in low-digit
milliseconds of clock skew, and we had few chances to reach larger
offsets. This adjusts the distribution to be flatter, giving us more
chances of picking large offsets.

* nemesis.time: disable systemd clock sync

Systemd took over from ntpd and has its own way to disable time
synchronization. Now we do both. Also changes offsets to be floats, not
bigdecimals.

* update rate-graph! to behave like point-graph!(latency-raw): elide empty [:f :type] values, use 1s bucket-times

* nemesis.combined/clock-package: fix final generator

The final generator did not actually reset the clocks!

* restore rate-graph! bucket-size to 10s

* Formatting

* util/zipf: generate Zipf-distributed integers

rand-distribution also supports zipfs now. Also adds data.generators as
a dependency.

* util/zipf: use zipf-default-skew

* util/rand-distribution: allow nil skew for zipfians

* nemesis.combined: fix misleading docstrings

The docs for packet and file corruption nemeses used :node-spec, not
node-spec.

* nemesis.combined: clean up unused binding

* nemesis/compose: better error messages for unknown `f`s

Passing an unknown :f to a ReflCompose nemesis would throw a cast
exception rather than the helpful error we intended.

* os/debian: add build-essential to default packages

The time nemesis needs this; it's silly that we installed it separately.

* nemesis.time: extract binary compiling, speed up disabling NTP

We're going to want to build C binaries in other nemeses, so the
compile-resource! functions now live in jepsen.nemesis. Also fixes an
issue which caused us to burn 30 seconds running tests on containers,
waiting for systemd's ntp-disabling command to have an internal network
timeout. JFC.

* nemesis: rename compile-resource to compile-c-resource etc

It seems likely we'll have other languages someday.

* nemesis.file: a new nemesis for filesystem corruption

I'd like to move some other things, like bitflip, in here too. Right now
there's one nemesis for file operations, which introduces corruption in
chunks of a file by copying from random other chunks. This is probably
going to destroy most databases.

* nemesis.file/corrupt-file-chunks-helix-gen: add an empty arity

* corrupt-file: fix incorrect argument parsing

* checker: fix a docstring

* corrupt-file: parse CLI options, allow start and end ranges

* nemesis.file: whoops, it's --modulus

* corrupt-file: print statistics

* corrupt-file: print final summary to stdout

* generator/concurrency-limit

A generator which limits the concurrency of running operations.

* util/minority: Like majority, returns less than half of n.

* nemesis/file: add snapshot/restore, rename corrupt->copy

This allows you to take snapshots of specific regions of a file, and
restore them later, simulating misdirected/lost writes. We also rename
:f :corrupt-file-chunks to :f :copy-file-chunks, because there are now
multiple senses of corruption.

* corrupt-file: better docs

* test: drop focus metadata

* corrupt-file: add support for --mode bitflip

* nemesis.file: fix reflection, lack of update in nemesis

* corrupt-file: don't run past --end when corrupting

* nemesis.file: add bitflips, make generators more generic

* corrupt-file: fix --index for --mode bitflip

* nemesis.file: add default arity for helix-gen

* nemesis.file/nodes-gen: flip arguments to be consistent with helix-gen

---------

Co-authored-by: Kyle Kingsbury <aphyr@aphyr.com>
Co-authored-by: Bin Wang <bin.wang@mail.binwang.me>
Co-authored-by: Yishuai Li <liyishuai.lys@alibaba-inc.com>
Co-authored-by: nurturenature <jeff.suttor@earthlink.net>