jepsen.core: before snarfing log files, try to kill DBs
It's somewhat inconvenient, especially with composed DBs, to try and
download log files when the DB is still running--the DB might move files
around as you're trying to tar or download them. If the DB supports it,
we now kill the DB just before downloading logs. --leave-db-running
skips this behavior--a little awkward, but that's mainly for debugging.