jepsen.control: with-node, with-nodes, with-nodes*, with-all-nodes
This has been bugging me for years. It's impossible to keep track of the
various macros for running things on one or more nodes, and I'd boxed
myself into a corner by using existing names in a way I couldn't change
without breaking backwards compatibility.

Now we have:

with-node: run body on a single node.
with-nodes: run body on selected nodes.
with-nodes*: run function on selected nodes.
with-all-nodes: run body on all nodes.

Later we'll deprecate on-nodes and with-test-nodes.